Extract LambdaArn validation logic
